flexilance-mvp/requirements.txt
Django==5.0.14
djangorestframework==3.15.1
Pillow==10.3.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
cloudinary==1.41.0
django-cloudinary-storage==0.3.0